+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ### 2.10 (unreleased)
 
+- added the "-engineCtrl" option to control hardware and CNG engines
+
 ### 2.9 (2024.06.29)
 
 - added a 64 bit long pseudo-random NONCE in the TSA request